Refactor the code to use differet approaches to play with the Ai WorkFlow
