+++ ai-providers/ai-workflow/src/main/java/dev/snowdrop/ai/workflow/NonAiAgentCreativeWriter.java
@@ -1,7 +1,6 @@
 package dev.snowdrop.ai.workflow;
 
 import dev.langchain4j.agentic.Agent;
-import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
 import io.quarkiverse.langchain4j.CreatedAware;
